backend/requirements.txt
Django==4.2
djangorestframework==3.14
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.3.0
mysqlclient==2.2
Pillow==10.0.0
markdown==3.5
bleach==6.0.0
python-dateutil==2.8.2
redis==4.6.0
